refactor ETD findings to have no static classes + override equals and hashcode
